Guard the address space, range-check operands, fix error line numbers
Address space (fixes #21):
- PROG_SIZE was 0xFFFF, one byte short of the address space, so an
  instruction assembled at .org 0xFFFF wrote out of bounds; it is now
  0x10000
- All output goes through emit_byte(), which refuses to wrap PC/DATA_PC
  past 0xFFFF and reports the overflow once instead of silently
  truncating the image
- Output length is tracked as prog_end (highest address written + 1), so
  a program ending exactly at 0xFFFF is written completely; ORG is
  range-checked via the new set_origin()

Operand ranges (completes #20):
- Values that do not fit the field the instruction encodes them in are
  reported instead of silently truncated, for instructions and for
  defb/defw alike

  The check only diagnoses, it never changes how many bytes are emitted.
  On pass 1 an expression like "defb end - start" evaluates against
  labels that are not known yet and can land far out of range; dropping
  those bytes moved every later label. monitor48k reproduced the
  reference ROM before this batch and does so again, which is what
  caught it.

Diagnostics (fixes #24):
- asm_load_buffer() reset asmlineno to 0 where flex counts from 1, so
  syntax errors were reported one line early; the NL rule now records the
  line that just ended, so semantic errors (which fire after the parser
  has read ahead) stay accurate too
- error_print() reports on pass 2 only, so nothing is printed twice;
  redefinitions are detected per pass via user_label::pass_defined
- ihex writing reports failure properly (save_array_to_ihex() returned -1
  on success) and terminates the EOF record; an unknown -x format is an
  error instead of silently producing an empty file

Testing: monitor48k.asm assembles to a byte-exact copy of the real 16K
Spectrum ROM shipped beside it, so wire that comparison in as a CTest
test (rom_monitor48k) rather than only checking the exit status.

ctest: 3/3 pass (rom_monitor48k, selftest_fallingblocks, opcode_suite
916/916).

+++ include/asmdasm/assembler.h
@@ -24,7 +24,7 @@ bool is_single(uint32_t opcode);
 bool is_prefixed(uint32_t opcode);
 bool is_double_prefixed(uint32_t opcode);
 int handle_instruction(char *instruction, intmax_t data, size_t size);
-void defw(uint16_t data);
+void defw(intmax_t data);
 void deft(char *text);
 void add_label(char *label, intmax_t address);
 intmax_t get_label_address(char *label);
